dev > main 통합 - 게시글 상세 조회 인증 권한 변경, 게시글 작성 /조회시 이미지 URL 추가 (#58)
* fix: user컨트롤러 여백 조정

* fix: 유저컨트롤러 여백 조정

* fix: 비회원에 대한 코드 수정

* feat: Joi를 이용한 유효성검사 추가

* feat: multer, aws s3Client npm install

* feat: upload middleware 구현

* fix: login미들웨어 추가 및 route 수정

* fix: 불필요한 코드 삭제

* feat: 게시글 작성 API -이미지 추가 저장하도록 구현

* fix: login미들웨어, route경로 수정, 주석 삭제

* fix: joi.helper에서 postId,commentId유효성검사 추가+ routes들에 추가

* fix: 이메일, 닉네임 중복 검사 작성

* fix: 전체 오류 수정

* fix: 중복검사 오타 수정

* feat: jwt 토큰 만료 시간 환경변수 처리

* feat: 누락된 업로드 미들웨어 추가, 게시글 상세조회에 fileUrl 추가

* fix:비회원 게시글 상세조회시 오류해결

---------

+++ api/repositories/likes.repository.js
@@ -5,18 +5,18 @@ class LikesRepository {
     return await Posts.findOne({ where: { postId } });
   };
 
-  like = async (userId, postId) => {
-    const findLike = await Likes.findOne({ where: { userId, postId } });
+  findLike = async (userId, postId) => {
+    return await Likes.findOne({ where: { userId, postId } });
+  };
 
-    if (findLike) {
-      Likes.destroy({
-        where: { userId: findLike.userId, postId: findLike.postId },
-      });
-      return "좋아요가 해제되었습니다";
-    } else {
-      Likes.create({ userId, postId });
-      return "좋아요가 등록되었습니다.";
-    }
+  deleteLike = async (userId, postId) => {
+    Likes.destroy({ where: { userId, postId } });
+    return;
+  };
+
+  createLike = async (userId, postId) => {
+    Likes.create({ userId, postId });
+    return;
   };
 }
